+++ src/ui/App.tsx
@@ -11,7 +11,7 @@ import AppActions from './actions/AppActions';
 import * as PlayerActions from './actions/PlayerActions';
 
 import * as styles from './App.module.css';
-import { isCtrlKey } from './utils/utils-platform';
+import { isCtrlKey } from './lib/utils-platform';
 
 /*
 |--------------------------------------------------------------------------
