+++ src/ui/actions/LibraryActions.ts
@@ -10,8 +10,8 @@ import store from '../store';
 import types from '../constants/action-types';
 
 import * as app from '../lib/app';
-import * as utils from '../utils/utils';
-import * as m3u from '../utils/utils-m3u';
+import * as utils from '../lib/utils';
+import * as m3u from '../lib/utils-m3u';
 import { SortBy, TrackModel } from '../../shared/types/interfaces';
 import { SUPPORTED_PLAYLISTS_EXTENSIONS, SUPPORTED_TRACKS_EXTENSIONS } from '../../shared/constants';
 import * as PlaylistsActions from './PlaylistsActions';
