+++ src/ui/components/PlayingBarInfo/PlayingBarInfo.tsx
@@ -5,7 +5,7 @@ import ButtonRepeat from '../PlayerOptionsButtons/ButtonRepeat';
 
 import * as PlayerActions from '../../actions/PlayerActions';
 import Player from '../../lib/player';
-import * as utils from '../../utils/utils';
+import * as utils from '../../lib/utils';
 import { TrackModel, Repeat } from '../../../shared/types/interfaces';
 
 import * as styles from './PlayingBarInfo.module.css';
