+++ src/ui/reducers/library.ts
@@ -1,7 +1,7 @@
 import types from '../constants/action-types';
 
 import { config } from '../lib/app';
-import * as utils from '../utils/utils';
+import * as utils from '../lib/utils';
 import { Action, TrackModel, SortBy, SortOrder } from '../../shared/types/interfaces';
 
 export interface LibrarySort {
