+++ src/ui/views/Library/Library.tsx
@@ -4,7 +4,7 @@ import { useSelector } from 'react-redux';
 
 import * as ViewMessage from '../../elements/ViewMessage/ViewMessage';
 import TracksList from '../../components/TracksList/TracksList';
-import { filterTracks, sortTracks } from '../../utils/utils-library';
+import { filterTracks, sortTracks } from '../../lib/utils-library';
 import SORT_ORDERS from '../../constants/sort-orders';
 import { RootState } from '../../reducers';
